+++ princessBride/test/byui/cit260/princessBride/control/SceneControlTest.java
@@ -112,66 +112,65 @@ public class SceneControlTest {
         System.out.println("calculateIngredientVolume");
         
         System.out.println("\tTest case #1");
-        double radius = 1.0;
-        double height = 2.0;
+        int radius = 1;
+        int height = 2;
         SceneControl instance = new SceneControl();
         double expResult = 0.00628;
         double result = instance.calculateIngredientVolume(radius, height);
         assertEquals(expResult, result, 0.00001);
         
-        
         System.out.println("\tTest case #2");
-        radius = -1.0;
-        height = 2.0;
+        radius = -1;
+        height = 2;
         expResult = -1.0;
         result = instance.calculateIngredientVolume(radius, height);
         assertEquals(expResult, result, 0.0);
         
         System.out.println("\tTest case #3");
-        radius = 11.0;
-        height = 2.0;
+        radius = 11;
+        height = 2;
         expResult = -1.0;
         result = instance.calculateIngredientVolume(radius, height);
         assertEquals(expResult, result, 0.0);
         
         System.out.println("\tTest case #4");
-        radius = 1.0;
-        height = -1.0;
+        radius = 1;
+        height = -1;
         expResult = -1.0;
         result = instance.calculateIngredientVolume(radius, height);
         assertEquals(expResult, result, 0.0);
         
         System.out.println("\tTest case #5");
-        radius = 1.0;
-        height = 11.0;
+        radius = 1;
+        height = 11;
         expResult = -1.0;
         result = instance.calculateIngredientVolume(radius, height);
         assertEquals(expResult, result, 0.0);
         
         System.out.println("\tTest case #6");
-        radius = 0.0;
-        height = 2.0;
+        radius = 0;
+        height = 2;
         expResult = 0.0;
         result = instance.calculateIngredientVolume(radius, height);
         assertEquals(expResult, result, 0.0);
         
         System.out.println("\tTest case #7");
-        radius = 10.0;
-        height = 2.0;
+        radius = 10;
+        height = 2;
         expResult = 0.6283;
         result = instance.calculateIngredientVolume(radius, height);
         assertEquals(expResult, result, 0.0001);
         
         System.out.println("\tTest case #8");
-        radius = 1.0;
-        height = 0.0;
+        radius = 1;
+        height = 0;
         expResult = 0.0;
         result = instance.calculateIngredientVolume(radius, height);
         assertEquals(expResult, result, 0.0001);
         
         System.out.println("\tTest case #9");
-        radius = 1.0;
-        height = 10.0;
+        radius = 1;
+        height = 10;
         expResult = 0.0314;
         result = instance.calculateIngredientVolume(radius, height);
         assertEquals(expResult, result, 0.0001);
